moving function to helper_func file
